requirements.txt
azure-identity
azure-storage-file-datalake
pandas
